refactor: 로또 생성 로직 도메인으로 분리
