+++ src/domain/generateLottos.js
@@ -1,3 +1,5 @@
+import Lotto from "../models/Lotto.js";
+
 export const generateLottoNumbers = () => {
   // 중복 없이 1~45 사이의 숫자 6개 랜덤 생성, 오름차순 정렬
   const lottoNumbers = [];
@@ -11,4 +13,9 @@ export const generateLottoNumbers = () => {
   }
 
   return lottoNumbers.sort((a, b) => a - b);
+};
+
+export const generateLottos = (price) => {
+  const count = price / 1000;
+  return Array.from({ length: count }, () => new Lotto(generateLottoNumbers()));
 };
